Timer is working Still need to Change some things

+++ Chest System/Assets/Scripts/Chest MVC/ChestModel.cs
@@ -8,7 +8,7 @@ public class ChestModel
     public int MaxCoins;
     public int MinGems;
     public int MaxGems;
-    public int ChestOpenTime;
+    public float ChestOpenTime;
     public int CoinRequiredToOpenChest;
     public int GemsRequiredToOpenChest;
